python/requirements.txt
msal
requests
PyYAML
